backend para revisar, README añadido, queda poco

+++ api/src/main/java/com/monteroantonio/PDAM23/controllers/ProductoController.java
@@ -59,6 +59,8 @@ public class ProductoController {
             @AuthenticationPrincipal User loggedUser,
             @PathVariable UUID id
             ){
+        if (service.findById(id) == null)
+            return ResponseEntity.noContent().build();
         service.deleteById(id);
         return ResponseEntity.noContent().build();
     }
